add base_module.py
