+++ src/kwja/models/typo_module.py
@@ -1,24 +1,21 @@
-import copy
 from statistics import mean
 from typing import Any, Dict, List, Optional, Tuple
 
 import hydra
-import pytorch_lightning as pl
 import torch
-from omegaconf import DictConfig, OmegaConf
+from omegaconf import DictConfig
 from transformers import PretrainedConfig, PreTrainedModel
 
 from kwja.evaluators.typo_module_metric import TypoModuleMetric
+from kwja.models.base_module import BaseModule
 from kwja.models.components.head import SequenceLabelingHead
 from kwja.utils.constants import RESOURCE_PATH
 from kwja.utils.loss import compute_token_mean_loss
-from kwja.utils.omegaconf import filter_dict_items
 
 
-class TypoModule(pl.LightningModule):
+class TypoModule(BaseModule):
     def __init__(self, hparams: DictConfig) -> None:
-        super().__init__()
-        self.save_hyperparameters(hparams)
+        super().__init__(hparams)
 
         if valid_corpora := getattr(hparams.datamodule, "valid", None):
             self.valid_corpora: List[str] = list(valid_corpora)
@@ -130,39 +127,3 @@ class TypoModule(pl.LightningModule):
             "ins_predictions": ins_predictions,
             "ins_probabilities": ins_max_probabilities,
         }
-
-    def configure_optimizers(self):
-        # Split weights in two groups, one with weight decay and the other not.
-        no_decay = ("bias", "LayerNorm.weight")
-        optimizer_grouped_parameters = [
-            {
-                "params": [
-                    p for n, p in self.named_parameters() if not any(nd in n for nd in no_decay) and p.requires_grad
-                ],
-                "weight_decay": self.hparams.optimizer.weight_decay,
-                "name": "decay",
-            },
-            {
-                "params": [
-                    p for n, p in self.named_parameters() if any(nd in n for nd in no_decay) and p.requires_grad
-                ],
-                "weight_decay": 0.0,
-                "name": "no_decay",
-            },
-        ]
-        optimizer = hydra.utils.instantiate(
-            self.hparams.optimizer, params=optimizer_grouped_parameters, _convert_="partial"
-        )
-
-        warmup_steps = self.hparams.warmup_steps
-        lr_scheduler = hydra.utils.instantiate(
-            self.hparams.scheduler, optimizer=optimizer, num_warmup_steps=warmup_steps
-        )
-        return {"optimizer": optimizer, "lr_scheduler": {"scheduler": lr_scheduler, "interval": "step", "frequency": 1}}
-
-    def on_save_checkpoint(self, checkpoint: Dict[str, Any]) -> None:
-        hparams: DictConfig = copy.deepcopy(checkpoint["hyper_parameters"])
-        OmegaConf.set_struct(hparams, False)
-        if self.hparams.ignore_hparams_on_save:
-            hparams = filter_dict_items(hparams, self.hparams.hparams_to_ignore_on_save)
-        checkpoint["hyper_parameters"] = hparams
